Day16 HTTP Methods
GET, POST, PUT, DELETE
